requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.5.3
pydantic[email]==2.5.3
python-multipart==0.0.6
